requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
pydantic[email]
dnspython
certifi
